// packages/core/src/types.ts
export interface PoiCategory {
  id: string;
  name: string;
  color: string;
  position: number;
}